+++ pkg/innerring/config/fee.go
@@ -8,9 +8,7 @@ import (
 // FeeConfig is an instance that returns extra fee values for contract
 // invocations without notary support.
 type FeeConfig struct {
-	registerNamedCnr,
-	mainchain,
-	sidechain fixedn.Fixed8
+	mainchain fixedn.Fixed8
 }
 
 // NewFeeConfig constructs FeeConfig from viper.Viper instance. Latter must not be nil.
@@ -18,21 +16,10 @@ type FeeConfig struct {
 // Fee for named container registration is taken from "fee.named_container_register" value.
 func NewFeeConfig(v *viper.Viper) *FeeConfig {
 	return &FeeConfig{
-		registerNamedCnr: fixedn.Fixed8(v.GetInt64("fee.named_container_register")),
-		mainchain:        fixedn.Fixed8(v.GetInt64("fee.main_chain")),
-		sidechain:        fixedn.Fixed8(v.GetInt64("fee.side_chain")),
+		mainchain: fixedn.Fixed8(v.GetInt64("fee.main_chain")),
 	}
 }
 
 func (f FeeConfig) MainChainFee() fixedn.Fixed8 {
 	return f.mainchain
-}
-
-func (f FeeConfig) SideChainFee() fixedn.Fixed8 {
-	return f.sidechain
-}
-
-// NamedContainerRegistrationFee returns additional GAS fee for named container registration in NeoFS network.
-func (f FeeConfig) NamedContainerRegistrationFee() fixedn.Fixed8 {
-	return f.registerNamedCnr
 }