+++ pkg/innerring/processors/netmap/processor.go
@@ -75,8 +75,6 @@ type (
 
 		nodeValidator NodeValidator
 
-		notaryDisabled bool
-
 		nodeStateSettings state.NetworkSettings
 	}
 
@@ -100,17 +98,13 @@ type (
 
 		NodeValidator NodeValidator
 
-		NotaryDisabled bool
-
 		NodeStateSettings state.NetworkSettings
 	}
 )
 
 const (
-	newEpochNotification        = "NewEpoch"
-	addPeerNotification         = "AddPeer"
-	updatePeerStateNotification = "UpdateState"
-	removeNodeNotification      = "RemoveNode"
+	newEpochNotification   = "NewEpoch"
+	removeNodeNotification = "RemoveNode"
 )
 
 // New creates network map contract processor instance.
@@ -169,8 +163,6 @@ func New(p *Params) (*Processor, error) {
 
 		nodeValidator: p.NodeValidator,
 
-		notaryDisabled: p.NotaryDisabled,
-
 		nodeStateSettings: p.NodeStateSettings,
 	}, nil
 }
@@ -195,20 +187,6 @@ func (np *Processor) ListenerNotificationParsers() []event.NotificationParserInf
 	p.SetParser(netmapEvent.ParseNewEpoch)
 	parsers = append(parsers, p)
 
-	if !np.notaryDisabled {
-		return parsers
-	}
-
-	// new peer event
-	p.SetType(addPeerNotification)
-	p.SetParser(netmapEvent.ParseAddPeer)
-	parsers = append(parsers, p)
-
-	// update peer event
-	p.SetType(updatePeerStateNotification)
-	p.SetParser(netmapEvent.ParseUpdatePeer)
-	parsers = append(parsers, p)
-
 	return parsers
 }
 
@@ -230,20 +208,6 @@ func (np *Processor) ListenerNotificationHandlers() []event.NotificationHandlerI
 	// new epoch handler
 	i.SetType(newEpochNotification)
 	i.SetHandler(np.handleNewEpoch)
-	handlers = append(handlers, i)
-
-	if !np.notaryDisabled {
-		return handlers
-	}
-
-	// new peer handler
-	i.SetType(addPeerNotification)
-	i.SetHandler(np.handleAddPeer)
-	handlers = append(handlers, i)
-
-	// update peer handler
-	i.SetType(updatePeerStateNotification)
-	i.SetHandler(np.handleUpdateState)
 	handlers = append(handlers, i)
 
 	return handlers