+++ pkg/morph/client/netmap/update_state.go
@@ -39,8 +39,8 @@ func (u *UpdatePeerPrm) SetMaintenance() {
 func (c *Client) UpdatePeerState(p UpdatePeerPrm) error {
 	method := updateStateMethod
 
-	if c.client.WithNotary() && c.client.IsAlpha() {
-		// In notary environments Alphabet must calls UpdateStateIR method instead of UpdateState.
+	if c.client.IsAlpha() {
+		// Alphabet nodes must call UpdateStateIR method instead of UpdateState.
 		// It differs from UpdateState only by name, so we can do this in the same form.
 		// See https://github.com/nspcc-dev/neofs-contract/issues/225.
 		method += "IR"